+++ include/jenny5_events.h
@@ -17,11 +17,11 @@
 class jenny5_event{
 public:
 	char type;
-	intptr_t param1, param2;
+	int param1, param2;
 	int time;
 
 public:
-	jenny5_event(char _type, intptr_t _param1, intptr_t _param2, int _time)
+	jenny5_event(char _type, int _param1, int _param2, int _time)
 	{
 		type = _type;
 		param1 = _param1;
